Cleaning up settings module
This is a clean-up of code I wrote about two years ago.

Cleaning up the Settings class and all its modules. This includes some
code refactoring and improving doc-strings.
Some minor variable re-namings, too.
Also removing the colors() method, as it is un-needed at this point. At
present, I was using it for a single color variable which is not meant
to be user-configurable.